1.6.1 updates
- Feature: Awaitables support for versions 2023.1 or newer
- Enhancement: New inspector icons
- Bug Fix: Namespace fixes
- Change: Made the StatusManager a component and removed the IStatus interface

+++ Editor/Scripts/PropertyDrawers/ModuleContainerDrawer.cs
@@ -9,15 +9,16 @@ namespace StatusEffects.Inspector
     [CustomPropertyDrawer(typeof(ModuleContainer))]
     public class ModuleContainerDrawer : PropertyDrawer
     {
-        private float _fieldSize = EditorGUIUtility.singleLineHeight;
-        private float _padding = EditorGUIUtility.standardVerticalSpacing;
-        private int _fieldCount = 1;
-
         private SerializedProperty _module;
         private SerializedProperty _moduleInstance;
 
-        private SerializedObject instance;
-        private SerializedProperty iterator;
+        private SerializedObject _instance;
+        private SerializedProperty _iterator;
+
+        private float _fieldSize = EditorGUIUtility.singleLineHeight;
+        private float _padding = EditorGUIUtility.standardVerticalSpacing;
+        private float _derivedPropertyHeight;
+        private float _derivedFieldCount;
 
         private Type _moduleInstanceType;
         private Attribute _attribute;
@@ -32,15 +33,6 @@ namespace StatusEffects.Inspector
                 GUI.color = color;
                 return;
             }
-            // Multi-editing is not allowed
-            if (property.hasMultipleDifferentValues)
-            {
-                Color color = GUI.color;
-                GUI.color = Color.yellow;
-                EditorGUI.LabelField(position, "Cannot multi-edit modules!");
-                GUI.color = color;
-                return;
-            }
 
             _module = property.FindPropertyRelative("module");
             _moduleInstance = property.FindPropertyRelative("moduleInstance");
@@ -51,27 +43,13 @@ namespace StatusEffects.Inspector
             position.y += _padding;
             // Draw base module reference
             EditorGUI.PropertyField(position, _module, GUIContent.none);
-            // If the module reference was removed destroy the instance
-            if (_module.objectReferenceValue == null && _moduleInstance.objectReferenceValue != null)
+            position.y += _fieldSize + _padding;
+
+            if (!_module.hasMultipleDifferentValues)
             {
-                ScriptableObject instance = _moduleInstance.objectReferenceValue as ScriptableObject;
-                AssetDatabase.RemoveObjectFromAsset(instance);
-                EditorUtility.SetDirty(property.serializedObject.targetObject);
-                AssetDatabase.SaveAssetIfDirty(property.serializedObject.targetObject);
-                UnityEngine.Object.DestroyImmediate(instance);
-            }
-
-            if (_module.objectReferenceValue != null)
-            {
-                _attribute = Attribute.GetCustomAttribute(_module.objectReferenceValue.GetType(), typeof(AttachModuleInstanceAttribute));
-
-                if (_attribute == null)
-                    return;
-
-                _moduleInstanceType = ((AttachModuleInstanceAttribute)_attribute).type;
-                // If the module reference was changed we add or destroy the instance, note that
-                // we also check in the StatusEffectDataEditor for when list items are removed.
-                if (_moduleInstance.objectReferenceValue != null && _moduleInstanceType != _moduleInstance.objectReferenceValue.GetType())
+                var sp = property.serializedObject.GetIterator();
+                // If the module reference was removed destroy the instance
+                if (_module.objectReferenceValue == null && _moduleInstance.objectReferenceValue != null)
                 {
                     ScriptableObject instance = _moduleInstance.objectReferenceValue as ScriptableObject;
                     AssetDatabase.RemoveObjectFromAsset(instance);
@@ -79,31 +57,51 @@ namespace StatusEffects.Inspector
                     AssetDatabase.SaveAssetIfDirty(property.serializedObject.targetObject);
                     UnityEngine.Object.DestroyImmediate(instance);
                 }
-                if (_moduleInstance.objectReferenceValue == null)
-                {
-                    ScriptableObject instance = ScriptableObject.CreateInstance(_moduleInstanceType);
-                    AssetDatabase.AddObjectToAsset(instance, property.serializedObject.targetObject as ScriptableObject);
-                    EditorUtility.SetDirty(property.serializedObject.targetObject);
-                    AssetDatabase.SaveAssetIfDirty(property.serializedObject.targetObject);
-                    _moduleInstance.SetUnderlyingValue(instance);
-                }
-                // Iterate through the module instance to display properties even if they are derived.
-                instance = new SerializedObject(_moduleInstance.objectReferenceValue as ModuleInstance);
-                instance.Update();
 
-                iterator = instance.GetIterator();
-                // Skip the script property
-                iterator.NextVisible(true);
-
-                while (iterator.NextVisible(false))
+                if (_module.objectReferenceValue != null)
                 {
-                    position.y += _fieldSize + _padding;
-                    EditorGUI.BeginChangeCheck();
-                    EditorGUI.PropertyField(position, iterator);
-                    if (EditorGUI.EndChangeCheck())
-                        instance.ApplyModifiedProperties();
+                    _attribute = Attribute.GetCustomAttribute(_module.objectReferenceValue.GetType(), typeof(AttachModuleInstanceAttribute));
+
+                    if (_attribute == null)
+                        return;
+
+                    _moduleInstanceType = ((AttachModuleInstanceAttribute)_attribute).type;
+                    // If the module reference was changed we add or destroy the instance, note that
+                    // we also check in the StatusEffectDataEditor for when list items are removed.
+                    if (_moduleInstance.objectReferenceValue != null && _moduleInstanceType != _moduleInstance.objectReferenceValue.GetType())
+                    {
+                        ScriptableObject instance = _moduleInstance.objectReferenceValue as ScriptableObject;
+                        AssetDatabase.RemoveObjectFromAsset(instance);
+                        EditorUtility.SetDirty(property.serializedObject.targetObject);
+                        AssetDatabase.SaveAssetIfDirty(property.serializedObject.targetObject);
+                        UnityEngine.Object.DestroyImmediate(instance);
+                    }
+                    if (_moduleInstance.objectReferenceValue == null)
+                    {
+                        ScriptableObject instance = ScriptableObject.CreateInstance(_moduleInstanceType);
+                        AssetDatabase.AddObjectToAsset(instance, property.serializedObject.targetObject as ScriptableObject);
+                        EditorUtility.SetDirty(property.serializedObject.targetObject);
+                        AssetDatabase.SaveAssetIfDirty(property.serializedObject.targetObject);
+                        _moduleInstance.SetUnderlyingValue(instance);
+                    }
+                    // Iterate through the module instance to display properties even if they are derived.
+                    _instance = new SerializedObject(_moduleInstance.objectReferenceValue as ModuleInstance);
+                    _instance.Update();
+
+                    _iterator = _instance.GetIterator();
+                    // Skip the script property
+                    _iterator.NextVisible(true);
+
+                    while (_iterator.NextVisible(false))
+                    {
+                        EditorGUI.BeginChangeCheck();
+                        EditorGUI.PropertyField(position, _iterator);
+                        if (EditorGUI.EndChangeCheck())
+                            _instance.ApplyModifiedProperties();
+                        position.y += EditorGUI.GetPropertyHeight(_iterator) + _padding;
+                    }
+                    _instance.Dispose();
                 }
-                instance.Dispose();
             }
 
             EditorGUI.EndProperty();
@@ -111,11 +109,31 @@ namespace StatusEffects.Inspector
 
         public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
         {
+            _module = property.FindPropertyRelative("module");
             _moduleInstance = property.FindPropertyRelative("moduleInstance");
-            if (property.hasMultipleDifferentValues || _moduleInstance.GetUnderlyingValue() == null)
+
+            if (_module.hasMultipleDifferentValues || _moduleInstance.objectReferenceValue == null)
                 return _fieldSize + _padding * 2;
-            _fieldCount = _moduleInstance.GetUnderlyingValue().GetType().GetFields(System.Reflection.BindingFlags.Instance | System.Reflection.BindingFlags.Public).Length + 1;
-            return (_fieldSize + _padding) * _fieldCount + _padding;
+
+            _derivedFieldCount = 0; 
+            _derivedPropertyHeight = 0;
+            // Iterate through the module instance to display properties even if they are derived.
+            _instance = new SerializedObject(_moduleInstance.objectReferenceValue as ModuleInstance);
+            _instance.Update();
+
+            _iterator = _instance.GetIterator();
+            // Skip the script property
+            _iterator.NextVisible(true);
+
+            while (_iterator.NextVisible(false))
+            {
+                _derivedFieldCount++;
+                _derivedPropertyHeight += EditorGUI.GetPropertyHeight(_iterator);
+            }
+            
+            _instance.Dispose();
+            
+            return _fieldSize + _padding * (_derivedFieldCount + 2) + _derivedPropertyHeight;
         }
     }
 }
